requirements_prod.txt
alembic~=1.9.4
aiohttp~=3.8.4
aiosmtplib~=2.0.1
algoliasearch~=3.0.0
anyio~=3.6.2
APScheduler~=3.10.0
asyncpg~=0.27.0
bcrypt~=4.0.1
fastapi~=0.89.1
Jinja2~=3.1.2
psycopg2-binary~=2.9.5
pydantic~=1.10.4
python-multipart~=0.0.5
PyJWT~=2.6.0
pyotp~=2.8.0
redis~=4.5.0
sqladmin~=0.9.0
SQLAlchemy~=2.0.0
uvicorn~=0.20.0
